+++ ZEA.Architectures.Mediators.MediatrWrapper.Sample/Requests/AddRequest.cs
@@ -1,0 +1,5 @@
+using ZEA.Architectures.Mediators.Abstractions.Interfaces;
+
+namespace ZEA.Architectures.Mediators.MediatrWrapper.Sample.Requests;
+
+public sealed record AddRequest(double SummandOne, double SummandTwo) : IRequest<double>;