+++ ZEA.Architectures.Mediators.MediatrWrapper.Sample/Requests/SubtractRequest.cs
@@ -1,0 +1,5 @@
+using ZEA.Architectures.Mediators.Abstractions.Interfaces;
+
+namespace ZEA.Architectures.Mediators.MediatrWrapper.Sample.Requests;
+
+public sealed record SubtractRequest(double Minuend, double Subtrahend) : IRequest<double>;