+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ---
 
+[2.2.2] - 2024.10.16
+
+### Fixed
+
+- **ZEA.Communications.Messaging.MassTransit.Generators.AzureServiceBus**
+    - Fixed a wrong namespace
+
+---
+
 ## [2.2.0] - 2024-09-24
 
 ### Added